requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
openai==2.26.0
pypandoc-binary==1.16.2
python-pptx==1.0.2
gTTS==2.5.4
moviepy==2.2.1
imageio-ffmpeg==0.6.0
pytest==9.0.2
